==> src/3-resampling.ipynb <==
import pandas as pd
from imblearn.over_sampling import SMOTE
from imblearn.under_sampling import RandomUnderSampler

dataset = pd.read_csv("../data/pca/pca_fingerprint_results.csv")
dataset

dataset["class"].value_counts()

# 1. Split Data between Ligand and Decoy compounds

ligand_df = dataset[dataset["class"] == 1].reset_index(drop=True)
decoy_df = dataset[dataset["class"] == 0].reset_index(drop=True)

ligand_df.head(10)

decoy_df.head(10)

ligand_df.info() # have 78 entries

decoy_df.info() # have 2322 entries

# 2. Take 150 random samples of decoy compounds

decoy_samples = decoy_df.sample(n=150, random_state=42).reset_index(drop=True) # Take 150 random samples of decoy compound
ligand_samples = ligand_df.copy() # Take all 78 samples of ligand compound

decoy_samples

ligand_samples

# 3. Concatenate data of randomly sampled Decoy with Ligand

sampled_data = pd.concat([decoy_samples, ligand_samples], axis=0).reset_index(drop=True)
sampled_data.head(10)

sampled_data.info() # Have 228 entries, 150 of decoys and 78 of ligands

# 4. Oversampling of Ligand using SMOTE

# Separate features and labels
X = sampled_data.drop(columns=['class'])
y = sampled_data['class']

# Initiate SMOTE
smote = SMOTE(sampling_strategy='auto', random_state=7)

# Do oversampling
X_resampled, y_resampled = smote.fit_resample(X, y)

# Create a new DataFrame with the oversampled data
oversampled_data = pd.concat([pd.DataFrame(X_resampled, columns=X.columns), pd.Series(y_resampled, name='class')], axis=1)
oversampled_data

oversampled_data.info() # have 300 entries

oversampled_data["class"].value_counts() # Balanced data between ligand and decoy

oversampled_data.to_csv("../data/resampled/oversampled_data.csv", index=False)

# 5. Undersampling of Decoy using RandomRandomSampler

# Initiate undersampler
undersampler = RandomUnderSampler(sampling_strategy='auto', random_state=3)

# Do undersamping
X_resampled, y_resampled = undersampler.fit_resample(X, y)

# Create a new DataFrame with the undersampled data
undersampled_data = pd.concat([pd.DataFrame(X_resampled, columns=X.columns), pd.Series(y_resampled, name='class')], axis=1).reset_index(drop=True)
undersampled_data.head(10)

undersampled_data.info()

undersampled_data["class"].value_counts() # Balanced data between ligand and decoy

undersampled_data.to_csv("../data/resampled/undersampled_data.csv", index=False)
